refactor: :recycle: archivo de entidad categoria movido a donde corresponde
se movio category.dart a la carpte de dominio del feature category
