+++ lib/features/category/data/categoryListScreen.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'package:flutter/material.dart';
-import 'package:godeliveryapp_naranja/entities/category.dart';
+import 'package:godeliveryapp_naranja/features/category/domain/category.dart';
 import 'package:godeliveryapp_naranja/features/category/presentation/widgets/category_card.dart';
 import 'package:http/http.dart' as http;
